Update file menu for the waveform, make new ColorDialog class
